Bugfix/bna demo for April 18th release (#617)
* [docker-compose] ad network configuration

Setting network configuration to allow integration with bidding and auction demo.
Privacy Sandbox Demos will use 172.16.128.0/24 range to not collide with Bidding and Auction services with fixed ips on 172.16.0.0/24

* package.json scripts

- update start command to be able to refresh a specific service. use `npm run start -- [service-name]`
- update stop command to remove volumes
- update build command to build without cache

* docker-compose : fix race condition on ad-tech services

ad-tech services are using the same codebase and were all mounting the local ./services/ad-tech folder causing race conditions. Often one of the ad-tech services would not start properly. as a temporary workaround we are removing the filesystem mounted volume.